stereoselectivity/requirements.txt
biopython==1.81
certifi==2023.7.22
charset-normalizer==3.2.0
dnachisel==3.2.11
docopt==0.6.2
flametree==0.1.12
idna==3.4
numpy==1.26.0
proglog==0.1.10
python-codon-tables==0.1.12
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.5
